Crate pallet_gear

Re-exports

pub use crate::manager::ExtManager;

pub use crate::manager::HandleKind;

pub use weights::WeightInfo;

pub use crate::pallet::*;

Modules

manager

Manager which handles results of message processing.

migrations

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

pallet_tests

weights

Autogenerated weights for pallet_gear

Macros

impl_config

impl_config_inner

Structs

GasInfo

The struct contains results of gas calculation required to process a message.

InstructionWeights

Describes the weight for all categories of supported wasm instructions.

Limits

Describes the upper limits on various metrics.

MemoryWeights

Describes the weight for memory interaction.

ReplyInfo

The struct contains results of read only send message RPC call.

Schedule

Definition of the cost schedule and other parameterization for the wasm vm.

SyscallWeights

Describes the weight for each imported function that a program is allowed to call.

Traits

BuiltinDispatcher

A trait representing a registry that provides methods to lookup and run a builtin actor.

BuiltinDispatcherFactory

A trait that defines the interface of a builtin dispatcher factory.

DebugInfo

Type Aliases

Authorship

BlockGasLimitOf

Ext

GasAllowanceOf

GasBalanceOf

GasHandlerOf

GasNodeIdOf

HandleFn

ProgramStorageOf

TaskPoolOf