Module pallet_gear::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

claim_value

Claim value from message in Mailbox.

create_program

Creates program via code_id from storage.

run

Process message queue

send_message

Sends a message to a program or to another account.

send_reply

Send reply on message in Mailbox.

set_execute_inherent

Sets ExecuteInherent flag.

upload_code

Saves program code in storage.

upload_program

Creates program initialization request (message), that is scheduled to be run in the same block.