pallet_gear::pallet::dispatchables

Function claim_value

pub fn claim_value<T: Config>(message_id: MessageId)

Claim value from message in Mailbox.

Removes message by given MessageId from callers Mailbox: rent funds become free, associated with the message value transfers from message sender to extrinsic caller.

NOTE: only user who is destination of the message, can claim value or reply on the message from mailbox.

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::claim_value.