pallet_gear::pallet

Module storage_types

Auto-generated docs-only module listing all (public and private) defined storage types for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes.

To see the actual storage type, find a struct with the same name at the root of the pallet, in the list of Type Definitions.

Structs

BenchmarkStorage

Warning: Doc-Only

BlockNumber

The current block number being processed.

ExecuteInherent

A flag indicating whether the message queue should be processed at the end of a block

GearRunInBlock

A guard to prohibit all but the first execution of pallet_gear::run() call in a block.