Struct pallet_gear::pallet::Pallet

pub struct Pallet<T>(_);

The pallet implementing the on-chain logic.

Implementations

impl<T: Config> Pallet<T>where T::AccountId: Origin,

pub fn set_block_number(bn: u32)

Set gear block number.

For tests only.

pub fn read_state_using_wasm( program_id: H256, fn_name: Vec<u8>, wasm: Vec<u8>, argument: Option<Vec<u8>> ) -> Result<Vec<u8>, Vec<u8>>

pub fn read_state(program_id: H256) -> Result<Vec<u8>, Vec<u8>>

pub fn read_metahash(program_id: H256) -> Result<H256, Vec<u8>>

pub fn calculate_gas_info( source: H256, kind: HandleKind, payload: Vec<u8>, value: u128, allow_other_panics: bool, initial_gas: Option<u64> ) -> Result<GasInfo, Vec<u8>>

pub fn run_with_ext_copy<R, F: FnOnce() -> R>(f: F) -> R

pub fn is_initialized(program_id: ProgramId) -> bool

Returns true if a program has been successfully initialized

pub fn is_active(program_id: ProgramId) -> bool

Returns true if id is a program and the program has active status.

pub fn is_terminated(program_id: ProgramId) -> bool

Returns true if id is a program and the program has terminated status.

pub fn is_exited(program_id: ProgramId) -> bool

Returns true if id is a program and the program has exited status.

pub fn get_block_number(program_id: ProgramId) -> <T as Config>::BlockNumber

pub fn exit_inheritor_of(program_id: ProgramId) -> Option<ProgramId>

Returns exit argument of an exited program.

pub fn termination_inheritor_of(program_id: ProgramId) -> Option<ProgramId>

Returns inheritor of terminated (failed it’s init) program.

pub fn next_message_id(user_id: H256) -> MessageId

Returns MessageId for newly created user message.

pub fn process_tasks(ext_manager: &mut ExtManager<T>)

Delayed tasks processing.

pub fn run_call() -> Call<T>

impl<T: Config> Pallet<T>where T::AccountId: Origin,

pub fn upload_code( origin: OriginFor<T>, code: Vec<u8> ) -> DispatchResultWithPostInfo

Saves program code in storage.

The extrinsic was created to provide deploy program from program functionality. Anyone who wants to define a “factory” logic in program should first store the code and metadata for the “child” program in storage. So the code for the child will be initialized by program initialization request only if it exists in storage.

More precisely, the code and its metadata are actually saved in the storage under the hash of the code. The code hash is computed as Blake256 hash. At the time of the call the code hash should not be in the storage. If it was stored previously, call will end up with an CodeAlreadyExists error. In this case user can be sure, that he can actually use the hash of his program’s code bytes to define “program factory” logic in his program.

Parameters

code: wasm code of a program as a byte vector.

Emits the following events:

SavedCode(H256) - when the code is saved in storage.

pub fn upload_program( origin: OriginFor<T>, code: Vec<u8>, salt: Vec<u8>, init_payload: Vec<u8>, gas_limit: u64, value: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance ) -> DispatchResultWithPostInfo

Creates program initialization request (message), that is scheduled to be run in the same block.

There are no guarantees that initialization message will be run in the same block due to block gas limit restrictions. For example, when it will be the message’s turn, required gas limit for it could be more than remaining block gas limit. Therefore, the message processing will be postponed until the next block.

ProgramId is computed as Blake256 hash of concatenated bytes of code + salt. (todo #512 code_hash + salt) Such ProgramId must not exist in the Program Storage at the time of this call.

There is the same guarantee here as in upload_code. That is, future program’s code and metadata are stored before message was added to the queue and processed.

The origin must be Signed and the sender must have sufficient funds to pay for gas and value (in case the latter is being transferred).

Parameters:

code: wasm code of a program as a byte vector.

salt: randomness term (a seed) to allow programs with identical code to be created independently.

init_payload: encoded parameters of the wasm module init function.

gas_limit: maximum amount of gas the program can spend before it is halted.

value: balance to be transferred to the program once it’s been created.

Emits the following events:

InitMessageEnqueued(MessageInfo) when init message is placed in the queue.

Note

Faulty (uninitialized) programs still have a valid addresses (program ids) that can deterministically be derived on the caller’s side upfront. It means that if messages are sent to such an address, they might still linger in the queue.

In order to mitigate the risk of users’ funds being sent to an address, where a valid program should have resided, while it’s not, such “failed-to-initialize” programs are not silently deleted from the program storage but rather marked as “ghost” programs. Ghost program can be removed by their original author via an explicit call. The funds stored by a ghost program will be release to the author once the program has been removed.

pub fn create_program( origin: OriginFor<T>, code_id: CodeId, salt: Vec<u8>, init_payload: Vec<u8>, gas_limit: u64, value: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance ) -> DispatchResultWithPostInfo

Creates program via code_id from storage.

Parameters:

code_id: wasm code id in the code storage.

salt: randomness term (a seed) to allow programs with identical code to be created independently.

init_payload: encoded parameters of the wasm module init function.

gas_limit: maximum amount of gas the program can spend before it is halted.

value: balance to be transferred to the program once it’s been created.

Emits the following events:

InitMessageEnqueued(MessageInfo) when init message is placed in the queue.

NOTE

For the details of this extrinsic, see upload_code.

pub fn send_message( origin: OriginFor<T>, destination: ProgramId, payload: Vec<u8>, gas_limit: u64, value: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance ) -> DispatchResultWithPostInfo

Sends a message to a program or to another account.

The origin must be Signed and the sender must have sufficient funds to pay for gas and value (in case the latter is being transferred).

To avoid an undefined behavior a check is made that the destination address is not a program in uninitialized state. If the opposite holds true, the message is not enqueued for processing.

Parameters:

destination: the message destination.

payload: in case of a program destination, parameters of the handle function.

gas_limit: maximum amount of gas the program can spend before it is halted.

value: balance to be transferred to the program once it’s been created.

Emits the following events:

DispatchMessageEnqueued(MessageInfo) when dispatch message is placed in the queue.

pub fn send_reply( origin: OriginFor<T>, reply_to_id: MessageId, payload: Vec<u8>, gas_limit: u64, value: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance ) -> DispatchResultWithPostInfo

Send reply on message in Mailbox.

Removes message by given MessageId from callers Mailbox: rent funds become free, associated with the message value transfers from message sender to extrinsic caller.

Generates reply on removed message with given parameters and pushes it in MessageQueue.

NOTE: source of the message in mailbox guaranteed to be a program.

NOTE: only user who is destination of the message, can claim value or reply on the message from mailbox.

pub fn claim_value( origin: OriginFor<T>, message_id: MessageId ) -> DispatchResultWithPostInfo

Claim value from message in Mailbox.

Removes message by given MessageId from callers Mailbox: rent funds become free, associated with the message value transfers from message sender to extrinsic caller.

NOTE: only user who is destination of the message, can claim value or reply on the message from mailbox.

pub fn run(origin: OriginFor<T>) -> DispatchResultWithPostInfo

Process message queue

pub fn set_execute_inherent(origin: OriginFor<T>, value: bool) -> DispatchResult

Sets ExecuteInherent flag.

Requires root origin (eventually, will only be set via referendum)

impl<T: Config> Pallet<T>

pub fn execute_inherent() -> bool

A flag indicating whether the message queue should be processed at the end of a block

If not set, the inherent extrinsic that processes the queue will keep throwing an error thereby making the block builder exclude it from the block.

impl<T: Config> Pallet<T>

pub fn block_number() -> T::BlockNumber

The current block number being processed.

It shows block number in which queue is processed. May be less than system pallet block number if panic occurred previously.

impl<T: Config> Pallet<T>

pub fn last_gear_block_number() -> Option<T::BlockNumber>

The Gear block number before processing messages.

A helper variable that mirrors the BlockNumber at the beginning of a block. Allows to gauge the actual BlockNumber progress.