Trait pallet_gear::BuiltinDispatcher

pub trait BuiltinDispatcher {
    type Error;

    // Required methods
    fn lookup<'a>(&'a self, id: &ProgramId) -> Option<&'a HandleFn<Self::Error>>;
    fn run(
        &self,
        f: &HandleFn<Self::Error>,
        dispatch: StoredDispatch,
        gas_limit: u64,
    ) -> Vec<JournalNote>;
}

A trait representing a registry that provides methods to lookup and run a builtin actor.

Required Associated Types

type Error

Required Methods

fn lookup<'a>(&'a self, id: &ProgramId) -> Option<&'a HandleFn<Self::Error>>

Looks up a builtin actor by its actor id.

fn run( &self, f: &HandleFn<Self::Error>, dispatch: StoredDispatch, gas_limit: u64, ) -> Vec<JournalNote>