Crate pallet_gear

Re-exports

pub use crate::manager::ExtManager;

pub use crate::manager::HandleKind;

pub use weights::WeightInfo;

pub use crate::pallet::*;

Modules

manager

Manager which handles results of message processing.

migrations

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

pallet_tests

weights

Autogenerated weights for pallet_gear

Macros

impl_config

impl_config_inner

Structs

BuiltinInfo

BuiltinReply

The result of a builtin actor handle call.

DbWeights

Describes DB access weights.

GasInfo

Pre-calculated gas consumption estimate for a message.

InstantiationWeights

Describes the weight for instantiation of the module.

InstructionWeights

Describes the weight for all categories of supported wasm instructions.

InstrumentationWeights

Describes WASM code instrumentation weights.

Limits

Describes the upper limits on various metrics.

MemoryWeights

Describes the weight for memory interaction.

RentWeights

Describes the weight for renting.

ReplyInfo

Pre-calculated reply information.

Schedule

Definition of the cost schedule and other parameterization for the wasm vm.

SyscallWeights

Describes the weight for each imported function that a program is allowed to call.

TaskWeights

Describes weights for running tasks.

Constants

EXISTENTIAL_DEPOSIT_LOCK_ID

Lock for programs on ED.

Traits

BuiltinDispatcher

A trait representing a registry that provides methods to lookup and run a builtin actor.

BuiltinDispatcherFactory

A trait that defines the interface of a builtin dispatcher factory.

Type Aliases

Authorship

BlockGasLimitOf

Ext

GasAllowanceOf

GasBalanceOf

GasHandlerOf

GasNodeIdOf

HandleFn

Builtin actor handle function signature.

ProgramStorageOf

TaskPoolOf

WeightFn

Builtin actor max_gas function signature.