Crate pallet_gear

Re-exports

pub use crate::manager::ExtManager;

pub use crate::manager::HandleKind;

pub use weights::WeightInfo;

pub use crate::pallet::*;

Modules

manager

Manager which handles results of message processing.

migration

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_gear

Structs

GasInfo

The struct contains results of gas calculation required to process a message.

HostFnWeights

Describes the weight for each imported function that a program is allowed to call.

InstructionWeights

Describes the weight for all categories of supported wasm instructions.

Limits

Describes the upper limits on various metrics.

Schedule

Definition of the cost schedule and other parameterization for the wasm vm.

Enums

Forcing

Mode of forcing message queue processing

ProcessStatus

Possible queue processing states.

Traits

DebugInfo

Type Definitions

Authorship

BlockGasLimitOf

GasAllowanceOf

GasHandlerOf

GasUnitOf