Module pallet_gear::pallet::dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet. Note that members of this module cannot be used directly and are only provided for documentation purposes.

Functions

claim_value

Claim value from message in Mailbox.

create_program

Creates program via code_id from storage.

pay_program_rent

Pay additional rent for the program.

resume_session_commit

Finishes the program resume session.

resume_session_init

Starts a resume session of the previously paused program.

resume_session_push

Appends memory pages to the resume session.

run

Process message queue

send_message

Sends a message to a program or to another account.

send_message_with_voucher

Sends a message to a program using pre-allocated funds.

send_reply

Send reply on message in Mailbox.

send_reply_with_voucher

Sends the reply to a message in Mailbox using pre-allocated funds.

set_execute_inherent

Sets ExecuteInherent flag.

upload_code

Saves program code in storage.

upload_program

Creates program initialization request (message), that is scheduled to be run in the same block.