pallet_gear::pallet

Struct Pallet

pub struct Pallet<T>(/* private fields */);

The Pallet struct, the main type that implements traits and standalone functions within the pallet.

Implementations

impl<T: Config> Pallet<T>
where T::AccountId: Origin,

pub fn spend_gas( to: Option<<T as Config>::AccountId>, id: impl Into<GasNodeIdOf<T>>, amount: GasBalanceOf<T>, )

impl<T: Config> Pallet<T>
where T::AccountId: Origin,

pub fn set_block_number(bn: BlockNumberFor<T>)

Set gear block number.

For tests only.

pub fn read_state_using_wasm( program_id: H256, payload: Vec<u8>, fn_name: Vec<u8>, wasm: Vec<u8>, argument: Option<Vec<u8>>, gas_allowance: Option<u64>, ) -> Result<Vec<u8>, Vec<u8>>

pub fn read_state( program_id: H256, payload: Vec<u8>, gas_allowance: Option<u64>, ) -> Result<Vec<u8>, Vec<u8>>

pub fn read_metahash( program_id: H256, gas_allowance: Option<u64>, ) -> Result<H256, Vec<u8>>

pub fn calculate_gas_info( source: H256, kind: HandleKind, payload: Vec<u8>, value: u128, allow_other_panics: bool, initial_gas: Option<u64>, gas_allowance: Option<u64>, ) -> Result<GasInfo, Vec<u8>>

pub fn calculate_reply_for_handle( origin: H256, destination: H256, payload: Vec<u8>, gas_limit: u64, value: u128, allowance_multiplier: u64, ) -> Result<ReplyInfo, Vec<u8>>

pub fn run_with_ext_copy<R, F: FnOnce() -> R>(f: F) -> R

pub fn is_initialized(program_id: ActorId) -> bool

Returns true if a program has been successfully initialized

pub fn is_active(builtins: &impl BuiltinDispatcher, program_id: ActorId) -> bool

Returns true if program_id is that of a in active status or the builtin actor.

pub fn is_terminated(program_id: ActorId) -> bool

Returns true if id is a program and the program has terminated status.

pub fn is_exited(program_id: ActorId) -> bool

Returns true if id is a program and the program has exited status.

pub fn program_exists( builtins: &impl BuiltinDispatcher, program_id: ActorId, ) -> bool

Returns true if there is a program with the specified program_id`` (it may be paused) or this program_id` belongs to the built-in actor.

pub fn first_inheritor_of(program_id: ActorId) -> Option<ActorId>

Returns inheritor of an exited/terminated program.

pub fn next_message_id(user_id: H256) -> MessageId

Returns MessageId for newly created user message.

pub fn process_tasks(ext_manager: &mut ExtManager<T>)

Delayed tasks processing.

pub fn run_call(max_gas: Option<GasBalanceOf<T>>) -> Call<T>

impl<T: Config> Pallet<T>
where T::AccountId: Origin,

pub fn upload_code( origin: OriginFor<T>, code: Vec<u8>, ) -> DispatchResultWithPostInfo

Saves program code in storage.

The extrinsic was created to provide deploy program from program functionality. Anyone who wants to define a “factory” logic in program should first store the code and metadata for the “child” program in storage. So the code for the child will be initialized by program initialization request only if it exists in storage.

More precisely, the code and its metadata are actually saved in the storage under the hash of the code. The code hash is computed as Blake256 hash. At the time of the call the code hash should not be in the storage. If it was stored previously, call will end up with an CodeAlreadyExists error. In this case user can be sure, that he can actually use the hash of his program’s code bytes to define “program factory” logic in his program.

Parameters

code: wasm code of a program as a byte vector.

Emits the following events:

SavedCode(H256) - when the code is saved in storage.

pub fn upload_program( origin: OriginFor<T>, code: Vec<u8>, salt: Vec<u8>, init_payload: Vec<u8>, gas_limit: u64, value: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, keep_alive: bool, ) -> DispatchResultWithPostInfo

Creates program initialization request (message), that is scheduled to be run in the same block.

There are no guarantees that initialization message will be run in the same block due to block gas limit restrictions. For example, when it will be the message’s turn, required gas limit for it could be more than remaining block gas limit. Therefore, the message processing will be postponed until the next block.

ActorId is computed as Blake256 hash of concatenated bytes of code + salt. (todo #512 code_hash + salt) Such ActorId must not exist in the Program Storage at the time of this call.

There is the same guarantee here as in upload_code. That is, future program’s code and metadata are stored before message was added to the queue and processed.

The origin must be Signed and the sender must have sufficient funds to pay for gas and value (in case the latter is being transferred).

Gear runtime guarantees that an active program always has an account to store value. If the underlying account management platform (e.g. Substrate’s System pallet) requires an existential deposit to keep an account alive, the related overhead is considered an extra cost related with a program instantiation and is charged to the program’s creator and is released back to the creator when the program is removed. In context of the above, the value parameter represents the so-called reducible balance a program should have at its disposal upon instantiation. It is not used to offset the existential deposit required for an account creation.

Parameters:

code: wasm code of a program as a byte vector.

salt: randomness term (a seed) to allow programs with identical code to be created independently.

init_payload: encoded parameters of the wasm module init function.

gas_limit: maximum amount of gas the program can spend before it is halted.

value: balance to be transferred to the program once it’s been created.

Emits the following events:

InitMessageEnqueued(MessageInfo) when init message is placed in the queue.

Note

Faulty (uninitialized) programs still have a valid addresses (program ids) that can deterministically be derived on the caller’s side upfront. It means that if messages are sent to such an address, they might still linger in the queue.

In order to mitigate the risk of users’ funds being sent to an address, where a valid program should have resided, while it’s not, such “failed-to-initialize” programs are not silently deleted from the program storage but rather marked as “ghost” programs. Ghost program can be removed by their original author via an explicit call. The funds stored by a ghost program will be release to the author once the program has been removed.

pub fn create_program( origin: OriginFor<T>, code_id: CodeId, salt: Vec<u8>, init_payload: Vec<u8>, gas_limit: u64, value: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, keep_alive: bool, ) -> DispatchResultWithPostInfo

Creates program via code_id from storage.

Parameters:

code_id: wasm code id in the code storage.

salt: randomness term (a seed) to allow programs with identical code to be created independently.

init_payload: encoded parameters of the wasm module init function.

gas_limit: maximum amount of gas the program can spend before it is halted.

value: balance to be transferred to the program once it’s been created.

Emits the following events:

InitMessageEnqueued(MessageInfo) when init message is placed in the queue.

NOTE

For the details of this extrinsic, see upload_code.

pub fn send_message( origin: OriginFor<T>, destination: ActorId, payload: Vec<u8>, gas_limit: u64, value: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, keep_alive: bool, ) -> DispatchResultWithPostInfo

Sends a message to a program or to another account.

The origin must be Signed and the sender must have sufficient funds to pay for gas and value (in case the latter is being transferred).

To avoid an undefined behavior a check is made that the destination address is not a program in uninitialized state. If the opposite holds true, the message is not enqueued for processing.

Parameters:

destination: the message destination.

payload: in case of a program destination, parameters of the handle function.

gas_limit: maximum amount of gas the program can spend before it is halted.

value: balance to be transferred to the program once it’s been created.

Emits the following events:

DispatchMessageEnqueued(MessageInfo) when dispatch message is placed in the queue.

pub fn send_reply( origin: OriginFor<T>, reply_to_id: MessageId, payload: Vec<u8>, gas_limit: u64, value: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, keep_alive: bool, ) -> DispatchResultWithPostInfo

Send reply on message in Mailbox.

Removes message by given MessageId from callers Mailbox: rent funds become free, associated with the message value transfers from message sender to extrinsic caller.

Generates reply on removed message with given parameters and pushes it in MessageQueue.

NOTE: source of the message in mailbox guaranteed to be a program.

NOTE: only user who is destination of the message, can claim value or reply on the message from mailbox.

pub fn claim_value( origin: OriginFor<T>, message_id: MessageId, ) -> DispatchResultWithPostInfo

Claim value from message in Mailbox.

Removes message by given MessageId from callers Mailbox: rent funds become free, associated with the message value transfers from message sender to extrinsic caller.

NOTE: only user who is destination of the message, can claim value or reply on the message from mailbox.

pub fn run( origin: OriginFor<T>, max_gas: Option<GasBalanceOf<T>>, ) -> DispatchResultWithPostInfo

Process message queue

pub fn set_execute_inherent(origin: OriginFor<T>, value: bool) -> DispatchResult

Sets ExecuteInherent flag.

Requires root origin (eventually, will only be set via referendum)

pub fn claim_value_to_inheritor( origin: OriginFor<T>, program_id: ActorId, depth: NonZero<u32>, ) -> DispatchResultWithPostInfo

Transfers value from chain of terminated or exited programs to its final inheritor.

depth parameter is how far to traverse to inheritor. A value of 10 is sufficient for most cases.

Example of chain

Program #1 exits (e.g `gr_exit syscall) with argument pointing to user. Balance of program #1 has been sent to user.

Program #2 exits with inheritor pointing to program #1. Balance of program #2 has been sent to exited program #1.

Program #3 exits with inheritor pointing to program #2 Balance of program #1 has been sent to exited program #2.

So chain of inheritors looks like: Program #3 -> Program #2 -> Program #1 -> User.

We have programs #1 and #2 with stuck value on their balances. The balances should’ve been transferred to user (final inheritor) according to the chain. But protocol doesn’t traverse the chain automatically, so user have to call this extrinsic.

impl<T: Config> Pallet<T>
where T::AccountId: Origin,

pub fn send_message_impl( origin: <T as Config>::AccountId, destination: ActorId, payload: Vec<u8>, gas_limit: u64, value: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, keep_alive: bool, gas_sponsor: Option<<T as Config>::AccountId>, ) -> DispatchResultWithPostInfo

Underlying implementation of GearPallet::send_message.

pub fn send_reply_impl( origin: <T as Config>::AccountId, reply_to_id: MessageId, payload: Vec<u8>, gas_limit: u64, value: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance, keep_alive: bool, gas_sponsor: Option<<T as Config>::AccountId>, ) -> DispatchResultWithPostInfo

Underlying implementation of GearPallet::send_reply.

pub fn upload_code_impl(code: Vec<u8>) -> DispatchResultWithPostInfo

Underlying implementation of GearPallet::upload_code.