Trait pallet_gear::pallet::Config

pub trait Config: Config + Config + Config {

    type RuntimeEvent: From<Event<Self>> + TryInto<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type Randomness: Randomness<Self::Hash, Self::BlockNumber>;
    type Currency: LockableCurrency<Self::AccountId> + ReservableCurrency<Self::AccountId>;
    type GasPrice: GasPrice<Balance = <<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type WeightInfo: WeightInfo;
    type Schedule: Get<Schedule<Self>>;
    type OutgoingLimit: Get<u32>;
    type DebugInfo: DebugInfo;
    type CodeStorage: CodeStorage;
    type ProgramStorage: PausedProgramStorage<BlockNumber = Self::BlockNumber, Error = DispatchError, AccountId = Self::AccountId>;
    type MailboxThreshold: Get<u64>;
    type ReservationsLimit: Get<u64>;
    type Messenger: Messenger<BlockNumber = Self::BlockNumber, Capacity = u32, OutputError = DispatchError, MailboxFirstKey = Self::AccountId, MailboxSecondKey = MessageId, MailboxedMessage = UserStoredMessage, QueuedDispatch = StoredDispatch, WaitlistFirstKey = ProgramId, WaitlistSecondKey = MessageId, WaitlistedMessage = StoredDispatch, DispatchStashKey = MessageId>;
    type GasProvider: GasProvider<ExternalOrigin = Self::AccountId, NodeId = GasNodeId<MessageId, ReservationId>, Balance = u64, Error = DispatchError>;
    type BlockLimiter: BlockLimiter<Balance = GasBalanceOf<Self>>;
    type Scheduler: Scheduler<BlockNumber = Self::BlockNumber, Cost = u64, Task = ScheduledTask<Self::AccountId>>;
    type QueueRunner: QueueRunner<Gas = GasBalanceOf<Self>>;
    type Voucher: PaymentVoucher<Self::AccountId, ProgramId, <<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance, VoucherId = Self::AccountId>;
    type ProgramRentFreePeriod: Get<BlockNumberFor<Self>>;
    type ProgramResumeMinimalRentPeriod: Get<BlockNumberFor<Self>>;
    type ProgramRentCostPerBlock: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type ProgramResumeSessionDuration: Get<BlockNumberFor<Self>>;

}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeEvent: From<Event<Self>> + TryInto<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type Randomness: Randomness<Self::Hash, Self::BlockNumber>

The generator used to supply randomness to contracts through seal_random

type Currency: LockableCurrency<Self::AccountId> + ReservableCurrency<Self::AccountId>

Balances management trait for gas/value migrations.

type GasPrice: GasPrice<Balance = <<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

Gas to Currency converter

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.

type Schedule: Get<Schedule<Self>>

Cost schedule and limits.

type OutgoingLimit: Get<u32>

The maximum amount of messages that can be produced in single run.

type DebugInfo: DebugInfo

type CodeStorage: CodeStorage

Implementation of a storage for program binary codes.

type ProgramStorage: PausedProgramStorage<BlockNumber = Self::BlockNumber, Error = DispatchError, AccountId = Self::AccountId>

Implementation of a storage for programs.

type MailboxThreshold: Get<u64>

The minimal gas amount for message to be inserted in mailbox.

This gas will be consuming as rent for storing and message will be available for reply or claim, once gas ends, message removes.

Messages with gas limit less than that minimum will not be added in mailbox, but will be seen in events.

type ReservationsLimit: Get<u64>

Amount of reservations can exist for 1 program.

type Messenger: Messenger<BlockNumber = Self::BlockNumber, Capacity = u32, OutputError = DispatchError, MailboxFirstKey = Self::AccountId, MailboxSecondKey = MessageId, MailboxedMessage = UserStoredMessage, QueuedDispatch = StoredDispatch, WaitlistFirstKey = ProgramId, WaitlistSecondKey = MessageId, WaitlistedMessage = StoredDispatch, DispatchStashKey = MessageId>

Messenger.

type GasProvider: GasProvider<ExternalOrigin = Self::AccountId, NodeId = GasNodeId<MessageId, ReservationId>, Balance = u64, Error = DispatchError>

Implementation of a ledger to account for gas creation and consumption

type BlockLimiter: BlockLimiter<Balance = GasBalanceOf<Self>>

Block limits.

type Scheduler: Scheduler<BlockNumber = Self::BlockNumber, Cost = u64, Task = ScheduledTask<Self::AccountId>>

Scheduler.

type QueueRunner: QueueRunner<Gas = GasBalanceOf<Self>>

Message Queue processing routing provider.

type Voucher: PaymentVoucher<Self::AccountId, ProgramId, <<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance, VoucherId = Self::AccountId>

Type that allows to check calller’s eligibility for using voucher for payment.

type ProgramRentFreePeriod: Get<BlockNumberFor<Self>>

The free of charge period of rent.

type ProgramResumeMinimalRentPeriod: Get<BlockNumberFor<Self>>

The minimal amount of blocks to resume.

type ProgramRentCostPerBlock: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

The program rent cost per block.

type ProgramResumeSessionDuration: Get<BlockNumberFor<Self>>

The amount of blocks for processing resume session.