pallet_gear

Trait BuiltinDispatcher

pub trait BuiltinDispatcher {
    type Context;
    type Error;

    // Required methods
    fn lookup<'a>(
        &'a self,
        id: &ActorId,
    ) -> Option<BuiltinInfo<'a, Self::Context, Self::Error>>;
    fn run(
        &self,
        context: BuiltinInfo<'_, Self::Context, Self::Error>,
        dispatch: StoredDispatch,
        gas_limit: u64,
    ) -> Vec<JournalNote>;
}

A trait representing a registry that provides methods to lookup and run a builtin actor.

Required Associated Types

type Context

type Error

Required Methods

fn lookup<'a>( &'a self, id: &ActorId, ) -> Option<BuiltinInfo<'a, Self::Context, Self::Error>>

Looks up a builtin actor by its actor id.

fn run( &self, context: BuiltinInfo<'_, Self::Context, Self::Error>, dispatch: StoredDispatch, gas_limit: u64, ) -> Vec<JournalNote>